Change datatypes

+++ WeatherApp/Models/WeatherModel.cs
@@ -30,7 +30,7 @@ namespace WeatherApp.Models
     public class Main
     {
         [JsonProperty("temp")]
-        public int Temp { get; set; }
+        public double Temp { get; set; }
 
         [JsonProperty("pressure")]
         public int Pressure { get; set; }
@@ -39,10 +39,10 @@ namespace WeatherApp.Models
         public int Humidity { get; set; }
 
         [JsonProperty("temp_min")]
-        public int TempMin { get; set; }
+        public double TempMin { get; set; }
 
         [JsonProperty("temp_max")]
-        public int TempMax { get; set; }
+        public double TempMax { get; set; }
     }
 
     public class Wind
@@ -51,7 +51,7 @@ namespace WeatherApp.Models
         public double Speed { get; set; }
 
         [JsonProperty("deg")]
-        public int Deg { get; set; }
+        public double Deg { get; set; }
     }
 
     public class Clouds
